common simulation routes
