+++ yaptide/routes/common_sim_routes.py
@@ -3,18 +3,18 @@ import logging
 from flask import request
 from flask_restful import Resource
 
-from marshmallow import Schema, ValidationError
+from marshmallow import Schema
 from marshmallow import fields
 
 from yaptide.persistence.database import db
-from yaptide.persistence.models import SimulationModel, EstimatorModel, PageModel, UserModel
+from yaptide.persistence.models import SimulationModel, EstimatorModel, PageModel, UserModel, InputModel
 
 from yaptide.routes.utils.decorators import requires_auth
 from yaptide.routes.utils.response_templates import yaptide_response
 from yaptide.routes.utils.utils import check_if_job_is_owned_and_exist
 
 
-class Results(Resource):
+class SimulationResults(Resource):
     """Class responsible for managing results"""
 
     @staticmethod
@@ -79,7 +79,7 @@ class Results(Resource):
     @requires_auth(is_refresh=False)
     def get(user: UserModel):
         """Method returning job status and results"""
-        schema = Results.APIParametersSchema()
+        schema = SimulationResults.APIParametersSchema()
         errors: dict[str, list[str]] = schema.validate(request.args)
         if errors:
             return yaptide_response(message="Wrong parameters", code=400, content=errors)
@@ -107,3 +107,35 @@ class Results(Resource):
             return yaptide_response(message=f"Results for job: {job_id}, results from db", code=200, content={"estimators": result_estimators})
 
         return yaptide_response(message="Results are unavailable", code=404)
+
+
+class SimulationInputs(Resource):
+    """Class responsible for returning simulation input"""
+
+    class APIParametersSchema(Schema):
+        """Class specifies API parameters"""
+
+        job_id = fields.String()
+
+    @staticmethod
+    @requires_auth(is_refresh=False)
+    def get(user: UserModel):
+        """Method returning simulation input"""
+        schema = SimulationInputs.APIParametersSchema()
+        errors: dict[str, list[str]] = schema.validate(request.args)
+        if errors:
+            return yaptide_response(message="Wrong parameters", code=400, content=errors)
+        param_dict: dict = schema.load(request.args)
+        job_id = param_dict['job_id']
+
+        is_owned, error_message, res_code = check_if_job_is_owned_and_exist(job_id=job_id, user=user)
+        if not is_owned:
+            return yaptide_response(message=error_message, code=res_code)
+
+        simulation: SimulationModel = db.session.query(SimulationModel).filter_by(job_id=job_id).first()
+
+        input: InputModel = db.session.query(InputModel).filter_by(simulation_id=simulation.id).first()
+        if not input:
+            return yaptide_response(message="Input of simulation is unavailable", code=404)
+
+        return yaptide_response(message="Input of simulation", code=200, content={"input": input.data})
